requirements.txt
altair==5.5.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
av==13.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
cdflib==1.3.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloup==3.0.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
curl_cffi==0.11.3
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
easyocr==1.7.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.20.0
fonttools==4.56.0
fqdn==1.5.1
frozendict==2.4.6
fsspec==2025.9.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
glcontext==3.0.0
graphviz==0.20.3
h11==0.14.0
h5py==3.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
isosurfaces==0.1.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lazy_loader==0.4
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
moderngl==5.12.0
moderngl-window==3.1.1
mpmath==1.3.0
multitasking==0.0.11
narwhals==1.33.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.2
notebook==7.3.3
notebook_shim==0.2.4
numpy==2.2.4
opencv-python-headless==4.12.0.88
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peewee==3.18.1
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pycairo==1.28.0
pyclipper==1.3.0.post6
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.19.1
pyparsing==3.2.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pywin32==306
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2025.10.23
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
screeninfo==0.8.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==78.1.0
shapely==2.1.2
six==1.17.0
skia-pathops==0.8.0.post2
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacepy==0.7.0
srt==3.5.3
stack-data==0.6.3
streamlit==1.44.0
svgelements==1.9.6
sympy==1.14.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.10.16
tinycss2==1.4.0
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.13
xgboost==3.0.0
yfinance==0.2.62
